Declarative config
