+++ frontend/public/index.html
@@ -70,7 +70,7 @@
   <script>
     document.getElementById('postButton').addEventListener('click', async () => {
       try {
-        const response = await axios.post('http://localhost:8000/publisher/message', {
+        const response = await axios.post('http://localhost:8000/publisher', {
           // Add your POST data here if needed
         });
