+++ go.work
@@ -1,0 +1,7 @@
+go 1.21.5
+
+use (
+	./frontend
+	./publisher
+	./subscriber
+)
